[ID changes] user now should work without the casting issue

+++ HappyJourneyAirline/Models/Ticket.cs
@@ -35,7 +35,7 @@ namespace HappyJourneyAirline.Models
                 TicketClassID = reader.GetInt32(4),
                 TicketStatusID = reader.GetInt32(5),
                 PaymentID = reader.GetInt32(6),
-                AgencyID = !reader.IsDBNull(7) ? reader.GetInt32(7) : (long?)null
+                AgencyID = !reader.IsDBNull(7) ? reader.GetInt64(7) : (long?)null
             });
         }
 
@@ -139,12 +139,12 @@ namespace HappyJourneyAirline.Models
             {
                 Id = reader.GetInt32(0),
                 FlightID = reader.GetInt32(1),
-                UserID = reader.GetInt32(2),
+                UserID = reader.GetInt64(2),
                 Seat = reader.GetString(3),
                 TicketClassID = reader.GetInt32(4),
                 TicketStatusID = reader.GetInt32(5),
                 PaymentID = reader.GetInt32(6),
-                AgencyID = !reader.IsDBNull(7) ? reader.GetInt32(7) : (long?)null
+                AgencyID = !reader.IsDBNull(7) ? reader.GetInt64(7) : (long?)null
             });
             return result.Count > 0 ? result[0] : null;
         }
@@ -165,12 +165,12 @@ namespace HappyJourneyAirline.Models
             {
                 Id = reader.GetInt32(0),
                 FlightID = reader.GetInt32(1),
-                UserID = reader.GetInt32(2),
+                UserID = reader.GetInt64(2),
                 Seat = reader.GetString(3),
                 TicketClassID = reader.GetInt32(4),
                 TicketStatusID = reader.GetInt32(5),
                 PaymentID = reader.GetInt32(6),
-                AgencyID = !reader.IsDBNull(7) ? reader.GetInt32(7) : (long?)null
+                AgencyID = !reader.IsDBNull(7) ? reader.GetInt64(7) : (long?)null
             });
         }
         // Fetch all tickets for a specific flight
@@ -188,12 +188,12 @@ namespace HappyJourneyAirline.Models
             {
                 Id = reader.GetInt32(0),
                 FlightID = reader.GetInt32(1),
-                UserID = reader.GetInt32(2),
+                UserID = reader.GetInt64(2),
                 Seat = reader.GetString(3),
                 TicketClassID = reader.GetInt32(4),
                 TicketStatusID = reader.GetInt32(5),
                 PaymentID = reader.GetInt32(6),
-                AgencyID = !reader.IsDBNull(7) ? reader.GetInt32(7) : (long?)null
+                AgencyID = !reader.IsDBNull(7) ? reader.GetInt64(7) : (long?)null
             });
         }
     }
